requirements.txt
# FastAPI Backend Dependencies
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pydantic==2.8.2
pydantic-settings==2.2.1

# Health Plan Agent Dependencies
openai==1.58.1
anthropic==0.34.0
click==8.1.7
rich==14.1.0
asyncio==3.4.3

# Database and Utilities
supabase==2.8.0
httpx==0.27.0
python-multipart==0.0.9

# Railway and Production
gunicorn==21.2.0
